gv2gxl writeHdr: fix use after free
This call to `free` looked like it was deallocating the earlier allocation
within this function. But at this point `bp` points to the return value of a
call to `addid`. This return value is a `name` member of an item stored in the
`idList` dictionary that is still in use.

The original allocated memory of `bp` was and still is leaked.