fix update of cluster outline periphery in fdp layout
98d00ae10931a636cc5ec480092b4a20f5156877 introduced an extra invisible
periphery representing the outline of a node and
b574f6d31e240e7f92a2c3fa5eb221a25a715253 introduced use of it in the
`poly_inside` function. For regular nodes, this periphery is set to
its final value already in `poly_init`, but for clusters it must be
updated to the actual cluster outline periphery during fdp layout in
the `setClustNodes` function which the latter commit above failed to
do.

Fixes https://gitlab.com/graphviz/graphviz/-/issues/2345.