common: remove unused 'elist_fastapp'
This definition is in a public header so this is technically an API break. But
it is unlikely any users are relying on this because it is typically used in
combination with `alloc_elist`, `elist_append`, and `free_list`, which contain
uses of other macros that are _not_ exposed like `ALLOC` and `N_NEW`.