dotgen: remove limit on number of 'samehead', 'sametail' attributes
There was a per-node limit of 5 unique `samehead`/`sametail` attributes. This
was arbitrary and seems to have been introduced to allow stack-allocating data
structures. By moving to a dynamic list for the `same_t` working data, this
limit can be removed entirely.

This change is fairly mechanical; replacing operations on raw arrays with
operations on generic lists. But note that extra management logic needed to be
introduced wrt clearing list elements and freeing lists. Each time around the
loop in `dot_sameports` the raw arrays were implicitly cleared by resetting
`n_samehead` and `n_sametail` to 0. And the arrays were implicitly freed by
returning from this function.

We can drop the `n` parameter to `sameedge` which was only used for error
reporting.

Some accesses need to use `same_list_at` instead of `same_list_get` because the
containing expression is passed to `elist_append` which is a macro and uses its
input as an lvalue.

Gitlab: closes #452
Reported-by: tamar_a_roulettes