dotgen: replace an 'elist' with a generic list
This removes a minor footgun, in that `elist_append` is a macro and it is not
obvious that it assigns to one of its parameters. This is explicit in the
generic list equivalent, `edge_list_append`, by the caller having to take the
address of their input.

This also squashes two compiler warnings because the generic list uses `size_t`
for sizes. It also removes the only use of `elist_fastapp`