fix: ignore CPP line directives claiming a line number greater than 'INT_MAX'
When using a C pre-processor line directive with a very large claimed line
number, the computation of the current line would wrap negatively, something
that is undefined behavior with respect to the C standard. This change causes
the lexer to ignore any such directives.

To soapbox for a moment… this issue was found by fuzzing and perhaps indicates
the debatable utility of fuzzing Graphviz. This issue is real: such input caused
Graphviz to diverge from its expected behavior. But what security boundary is
being violated here? I (sadly) am not omniscient, but I cannot see how the
capability to force misreported line numbers is a primitive one can leverage
into escalation across a safety or security boundary. It is possible some wizard
out there has written

  if (graphviz says line number is negative) launch missiles

but one imagines there is likely a lower effort way of compromising the wizard’s
system. I say this partly as a brief explanation of the Graphviz maintainers’
lukewarm reaction to offers to fuzz Graphviz at an industrial scale. Your author
is a self-professed fuzzing fanboy, but would like to see some effort/techniques
towards quantifying the severity of fuzzing-found results.

Gitlab: fixes #1318
Reported-by: Google Autofuzz project