temporarily disable Pylint indentation checking
Black (which is about to be run on Graphviz Python sources) uses 4-space
indentation. We would like to do this transition incrementally, moving from the
current 2-space indentation to 4-space. To allow this, we need to stop Pylint
checking indentation for now. This will be re-enabled when the transition is
complete.

Gitlab: #2347