doc: update libguide.pdf to match sources for 'GVRENDER_*' constants
This table, and the rest of this document, was very out of date. This change
addresses specifically this table. The rest of the document remains out of date
and is most probably not used by contemporary programmers developing
applications using Graphviz libraries, who are better off reading the header
comments.

Gitlab: fixes #684
Reported-by: Steve Roush