common parse_layers: fix: replace layer accumulation logic with a generic list
This logic had an off by one error and, when parsing 128 or more layers, would
not allocate enough memory to fit them all. The strange design of book-ending
the layer list with a null pointer at each end should probably be re-examined in
future, but for now we just do the minimally invasive fix.

Gitlab: fixes #2355