common selectedLayer: replace an 'agxbuf' with 'gv_strdup'
Using an `agxbuf` here was overkill. The string being constructed is just a copy
of `spec`, which we can write simpler.

Ideally we would avoid copying altogether and operate on the original string
using lib/cgraph/tokenize.h. But `selectedLayer` involves two levels of
`strtok_r` and the lib/cgraph/tokenize.h API is not setup to tokenize from a
`strview_t` input instead of `char*`.