smyrna on_gvprbuttonload_clicked: fix incorrect agxbuf initialization
The underlying buffer that was being used for this agxbuf was `BUFSIZ` bytes,
but the call to `agxbinit` was claiming it was `SMALLBUF` bytes. `SMALLBUF` is a
constant defined within Graphviz to 128. `BUFSIZ` is a libc-provided constant
with a platform-dependent value. These are not guaranteed to have the same
value. If `BUFSIZ < SMALLBUF` loading a file larger than `BUFSIZ - 1` bytes
would result in an out-of-bounds write.

This change removes the backing buffer altogether, resolving this issue. The
agxbuf is used to store lines read from a user-provided file. The previous code
that optimized for files <128 bytes is a de-optimization in contemporary
environments.