circogen clone_graph: replace a raw buffer with an agxbuf
This avoids the mental load of having to check what `SMALLBUF` is and ensure the
printed string will fit within it. An agxbuf expands on-demand to fit its
contents. In the current implementation, this change incurs no heap allocation
on 64-bit platforms. The printed string here fits entirely inline in an agxbuf.

Gitlab: related to #1950