common: remove an agxbuf backing buffer and push the buffer inside 'global_def'
The string printed into this buffer is based on user input. So there is no
reason to suspect it will fit in `SMALLBUF` characters. Removing the backing
buffer simplifies this code and lets the agxbuf allocate and then clean up its
own backing memory.

Pushing this buffer inside the callee makes its lifetime and purpose clearer.