API BREAK: fix: make `sflag` and `eflag` 32-bit unsigned values
An arrow type is packed into 8 bits (see `ARR_TYPE_*` and `ARR_MOD_*`
constants). An array of 4 of these is packed into an “arrow flag,” values of
which are stored in various `flag` and `type` fields including `bezier.sflag`
and `bezier.eflag`.

There were two main problems with this:

  1. Packing 4 8-bit values into an `int` assumes that the `int` type is ≥ 32
     bits.

  2. On a platform with a 32-bit `int` (most of the currently supported
     platforms), the operations to set the last entry of the 4-entry array
     involve a shift into the sign bit, something that is undefined behavior in
     C.

This change avoids both of these problems by making it explicit that the 4-entry
array should be an unsigned 32-bit scalar.

Gitlab: related to #2294