wrap Catch2 unit tests to enable both JUnit and stderr output
Unfortunately, as explained in the comment in the Python script in this commit,
there appears to be no way to instruct a Catch2 binary to output _both_ JUnit
format to a file and diagnostic failures to stderr.¹ First running the test
binary in JUnit mode and then, if it fails, re-running it in console mode seems
a simple way around this. Unfortunately this does not seem straightforward to
achieve in CMake (`add_test` does not support multiple `COMMAND`s and scripting
something involving the `||` shell operator in a cross-platform way seems
complex).

This commit introduces a Python wrapper to perform the re-run logic. It is not
ideal, but seems the least complex way of achieving this.

Gitlab: closes #2358

¹ See
  https://stackoverflow.com/questions/56244534/catch2-output-to-console-and-file
  for some more information.