common: use a distinct type for arrow flags
Flags for arrows (type of the arrow and various modification bits) were being
stored in a `uint32_t` at known offsets. This had various downsides, including:

  1. Reading and writing fields of this (latent) struct required using bitwise
     operators. This is both error prone and permitted ambiguity in which
     `uint32_t` variables were storing a single flag and which were storing an
     array of flags.

  2. Using a scalar in this way defeats any type safety analysis the compiler
     can provide. It cannot determine where the boundaries between fields lie so
     cannot diagnose, e.g., writing `ARR_MOD_INV` to the type field instead of
     bit 5.

This change defines a proper struct for both an individual flag and for an array
of 4 flags. The actual bit layouts of both of these are intended to be
unchanged.

The definition of `arrowflag_t` in this commit uses a non-standard extension of
a `uint8_t` bitfield. This is technically not allowed by C99 but all major
compilers appear to support this extension.

Arrow flags outside of arrows.c are still represented as opaque `uint32_t`
values, which accurately models that no other code should need to deal with the
internals of this type.

Gitlab: closes #2294