pack push: abort on resource exhaustion instead of graceful failure
This was the sole call to `stack_push` in the code base, a function that
attempts to fail gracefully on arithmetic overflow or out of memory. This is
reasonable as a goal, except that tracing back up the call stack reveals that
`neato_layout` interprets this not as failure but as no connected components
found. Trying to handle these scenarios is not useful when the caller interprets
the result like this. Lets revert to exiting on these kind of failures, which
are already fail-stop in other code on this same call graph anyway.