cgraph: rename 'stack_push_or_exit' to 'stack_push'
This function was originally named to call out that it aborts execution on
failure. In the meantime, numerous Graphviz functions have appeared that abort
like this (e.g. `gv_alloc` and friends) without “or exit” in their names. It
seems reasonable now to abbreviate the name of this function too, with the
understanding that gracefully recovering from out-of-memory situations is no
longer a goal.