topfish smooth_vecs: remove reuse of 'smoothed_vec'
This function uses none of the previous data in the `smoothed_vec` array, so we
can shrink its lifetime by allocating it anew each time. This makes code clearer
to both readers and the compiler. At the same time we migrate this to use the
newer cgraph memory allocation wrappers.