sfdpgen beautify_leaves: fix: remove reuse of loop variables
Looking at this function as a whole, it seems clear the author did not intend to
reuse the outer loop variable `i` in the inner loop. The effect of this mistake
was either (1) some dimensions of the input matrix were skipped or (2) the outer
iteration became an infinite loop and Graphviz would never terminate.

This change has no effect on the test suite, which suggests there is no test
coverage for either of the above scenarios.

Gitlab: #2283