add a test case for 'beautify=true' problem
We need to disable this for Python 3.6 because the `.='text'` XPath feature used
in this test case was only introduced in Python 3.7. Python 3.6 is already
EOLed, so it is reasonable to wonder why a work around for it is needed.
Unfortunately we still (for another couple of weeks) support Ubuntu 18.04 which
has only Python 3.6.

Gitlab: #2283