sfdpgen beautify_leaves: correct padding calculation and make it apply
The padding computed here was far too large. It seems to have been derived based
on the follow on computation of the inter-node spacing, but unfortunately that
too is wrong. These two bugs were further masked by
ba776ee9271f6af1efec836d4193a621b65b9852 making all these steps a no-op.

This change corrects the padding to something reasonable and swaps the order of
statements to ensure it is actually applied. An upcoming commit will fix the
follow on computation of the inter-node spacing.

While making this change, we also improve some formatting and remove the now
unused `maxang`.

Gitlab: #2283