CI: fix: have windows_build.py write to log file directly instead of via pipe
The Windows builds (not Cygwin or MinGW, just native Windows) have been silently
failing since 9f4a4fc9cad641bf53fa8da93e0f12ce9d23202c. Compilation itself
encounters an error, but the job is marked as completing successfully. It seems
the shell piping in use here is masking/suppressing the exit status of
windows_build.py.

This is surprising as this has not been the behavior in the past – compilation
errors would correctly fail the job – and I am not aware of any relevant changes
to ci/* or .gitlab-ci.yml. My guess is that PowerShell in the Windows runners
has been silently upgraded and we are seeing the effects of this clause in the
documentation:¹

  Beginning in PowerShell 7.2, error records redirected from native commands,
  like when using redirection operators (`2>&1`), aren't written to the `$Error`
  variable and the preference variable `$ErrorActionPreference` doesn't affect
  the redirected output.

By avoiding shell redirection and instead doing this in Python, we should be
able to gain more stable environmental behavior.

¹ https://learn.microsoft.com/en-us/powershell/module/microsoft.powershell.core/about/about_preference_variables?view=powershell-7.3#erroractionpreference