dotgen class2: fix: use an ignored-edge-aware alternative to 'agfindedge'
When processing routing, a previously concentrated edge that was now marked
`IGNORED` could be found. This would result in erroring out when the edge was
later discovered to be not `NORMAL`.

Note that we also need to update the test case for #827 as it makes Graphviz now
able to successfully handle this graph.

Gitlab: #827, fixes #2087
Reported-by: Moritz Lenz
Suggested-by: M Funk