fix: recognize 'dot' output format as textual during testing
The #2089 test cases were written prior to the `dot` helper function being
introduced in the test suite. 297ed0bc7b61b2a6616c335726334836c8ffcc10 that
introduced it inadvertently made `test_2089` fail for a different reason than
originally: the subprocess output was returned as a `bytes` and then tried to
compare against a `str`. Following this change, `test_2089` still fails, but in
the way it was originally intended to, indicating #2089 has still not been
fixed.

Gitlab: #2089