dotgen make_leafslots: avoid reallocing rank collection when shuffling
It is not clear that the shuffling loop in this function was safe; it is moving
through the list backwards, but when copying each entry it is not obvious that
`ND_order(v) >= i`. Without this property, the copying loop would overwrite
entries it had yet to get to.

Trying to experiment with variations using the test suite indicates this code
has no test coverage.

This commit refactors the expansion into something that does not rely on the
aforementioned property to be true. While making this change we also move to the
newer allocation wrappers.