Autotools: remove support for generating Io language bindings
Io is a language created in 2002¹ for which preliminary Graphviz binding support
was introduced in 753b4dbddf3501169ea4c0351c556b88ecb7395d in 2005. The
anticipated SWIG support of the `-io` flag never arrived and I think it is safe
to say at this point it never will. This commit removes this build system code
that has never been functional.

¹ https://en.wikipedia.org/wiki/Io_(programming_language)