fix: when parsing user shapes, see XML declaration end before seeking SVG start
When parsing either an SVG file with its content all in one line:

  <?xml version="1.0" encoding="UTF-8"?><svg…

or one where the XML declaration itself spanned multiple lines:

  <?xml
  version="1.0"
  encoding="UTF-8"?>
  <svg…

the file would not correctly be identified as an SVG image. This change now
ensures we have reached the end of the XML declaration (the closing `>`) before
looking for the opening of the SVG definition (the `<svg`). This added logic is
not very efficient. But it is only significantly costly when the input file is
malformed or has an unusually long XML declaration.

This fragile, text-based logic should be replaced with a proper XML parsing
library in future.

Gitlab: fixes #2352
Reported-by: Peter