fix: when parsing user shapes, use a sliding window to recognize SVG start
This removes a prior assumption that the opening SVG tag (`<svg`) occurs at the
start of the first line after the XML declaration. Intervening white space no
longer causes an SVG to go unrecognized. Variants like `< svg` and `<SVG` will
still evade this algorithm, but it is at least an improvement on the current
state.

This added logic is not very efficient. But it is only significantly costly when
the input file is malformed or has an unusually long prelude.

This fragile, text-based logic should be replaced with a proper XML parsing
library in future.

Gitlab: #2352
Reported-by: Peter