update regression test for #144
The existing test logic focused only on the straight A->B edge
of the 144_[no_]ortho.dot graphs, that exhibited the head/tail swap.

This commit simplifies the existing assertion slightly and adds new asserts
that ensure that the angular A->C edge, which is routed in a zigzag line,
will have its head and tail arrows correctly placed as well.

The 144_ortho.dot and 144_no_ortho.dot are also prettified slightly. In
particular, "vee" and "inv" are used as arrow head and tail instead of
"dot" and "curve". These make the resulting graph images more readable
and result in polygons ("p") whose locations ("points") are now tested
directly, unlike before, where the placement of the labels was tested.

Gitlab: #106, #144